(feat) sdk/rust: api changes
